+++ flutter_local_notifications/lib/src/platform_specifics/android/styles/big_picture_style_information.dart
@@ -15,6 +15,7 @@ class BigPictureStyleInformation extends DefaultStyleInformation {
     bool htmlFormatContent = false,
     bool htmlFormatTitle = false,
     this.hideExpandedLargeIcon = false,
+    this.showBigPictureWhenCollapsed = false,
   }) : super(htmlFormatContent, htmlFormatTitle);
 
   /// Overrides ContentTitle in the big form of the template.
@@ -41,4 +42,9 @@ class BigPictureStyleInformation extends DefaultStyleInformation {
 
   /// Hides the large icon when showing the expanded notification.
   final bool hideExpandedLargeIcon;
+
+  /// Shows the big picture as a thumbnail in the collapsed notification.
+  ///
+  /// This property is only applicable to Android versions 12.0 or newer.
+  final bool showBigPictureWhenCollapsed;
 }