+++ flutter_local_notifications/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+
+* [Android] added support for `showBigPictureWhenCollapsed` in `BigPictureStyleInformation`. Thanks to the PR from [hiimax (Codematic)](https://github.com/hiimax)
+
 ## [22.0.1]
 
 * [Windows] Suppress warning around usage of experimental coroutines. This is to fix issue [#2777](https://github.com/MaikuB/flutter_local_notifications/issues/2777)